requirements.txt
asgiref==3.8.1
Django==5.1.4
django-admin-cli==0.1.1
djangorestframework==3.15.2
psycopg2-binary==2.9.10
sqlparse==0.5.3
